+++ src/order.ts
@@ -242,8 +242,6 @@ export type ITicketPriceSpecification = Omit<ICompoundPriceSpecification<ITicket
 export type IOfferOptimized4acceptedOffer = Pick<
     IOffer,
     'typeOf' | 'id' | 'itemOffered' | 'offeredThrough' | 'name'
-// 不要なので廃止(2023-07-01~)
-// | 'priceCurrency'
 >;
 /**
  * 受け入れオファー
@@ -253,12 +251,11 @@ export interface IAcceptedOffer<T extends IItemOffered> extends IOfferOptimized4
      * オファー対象アイテム
      */
     itemOffered: T;
-    // 不要なので廃止(2023-07-01~)
-    // seller: {
-    //     typeOf: OrganizationType.Corporation;
-    //     name?: string | IMultilingualString;
-    // };
     priceSpecification?: ITicketPriceSpecification;
+    /**
+     * 資産取引を特定する番号(2024-01-30~)
+     */
+    serialNumber?: string;
 }
 /**
  * 販売者