+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 注文取引にignoreAccpetedOffersFromResultオプションを追加
 - 注文取引検索条件拡張
 - 決済サービスプロバイダーのIPaymentUrlSettingsにuse3DSを追加
+- 注文オファーにserialNumberを追加
+- オファー承認アクションにinstrument.transactionNumberを保証
 
 ### Changed